fix flow director flex bytes on i40e
well, it's actually a work-around and not a fix
the new filter API unfortunately doesn't allow ignoring
ip addresses and udp ports (this is probably a bug in the i40e driver)

ixgbe has the same problem to some degree, it used to be possible to
install a filter saying "match if this is a UDP packet with these two
bytes in the payload, ignore IPs and ports". the UDP part is no longer
possible since DPDK 2.2; but matching IP without checking addresses still
works.

the i40e driver doesn't seem to allow this, and just L2_PAYLOAD doesn't
match the packet because it's UDP...